Rename process to output
