+++ src/utils/dolphin/index.ts
@@ -1,3 +1,3 @@
 export * from "./launcher";
-export * from "./process";
+export * from "./output";
 export * from "./queue";
